@@ -33,24 +33,24 @@
   },
   "homepage": "https://github.com/do-community/hub-for-good-list#readme",
   "dependencies": {
-    "@babel/runtime": "^7.17.0",
-    "do-bulma": "git+https://github.com/do-community/do-bulma.git",
-    "do-vue": "git+https://github.com/do-community/do-vue.git",
+    "@babel/runtime": "^7.17.2",
+    "do-bulma": "git+https://github.com/do-community/do-bulma.git#MattIPv4/vue-3",
+    "do-vue": "git+https://github.com/do-community/do-vue.git#MattIPv4/vue-3",
     "google-spreadsheet": "^3.2.0",
     "linkify-it": "^3.0.3",
-    "tlds": "^1.229.0",
-    "vue": "^2.6.14",
-    "vue-select": "^3.16.0"
+    "tlds": "^1.230.0",
+    "vue": "^3.2.31",
+    "vue-select": "^4.0.0-beta.3"
   },
   "devDependencies": {
     "@babel/eslint-parser": "^7.17.0",
     "copyfiles": "^2.4.1",
-    "eslint": "^8.8.0",
-    "eslint-plugin-vue": "^8.4.0",
-    "posthtml": "^0.16.5",
+    "eslint": "^8.10.0",
+    "eslint-plugin-vue": "^8.5.0",
+    "posthtml": "^0.16.6",
     "posthtml-extend": "^0.6.2",
-    "sass": "^1.49.7",
-    "stylelint": "^14.3.0",
+    "sass": "^1.49.9",
+    "stylelint": "^14.5.3",
     "stylelint-config-standard-scss": "^3.0.0",
     "stylelint-order": "^5.0.0",
     "vue-template-compiler": "^2.6.14"
